requirements.txt
pyanybar
ping3